IGAPP-178: Refactor CategoriesPage by moving components out
